AppRouter.tsx: add GameGuard.tsx to assure that we have currentGame in the store
- fetch the game id from the url with useParams
- When no game found -> show no game found
- Show Room, GameView or GameSummary depending on the game state
- Always show the "Loading..." screen for at least 500 ms when
the game needs to be fetched.

Rename Game to GameView because of the name clash with the model Game.
Delete unused GameRouter.tsx.
Remove all the polling for the current game from the components,
because the GameGuard does that now.

This simultaneously implements that when the game state changes,
all players switch to the corresponding view.

Issue: https://github.com/sopra-fs22-group-36/screw-your-neighbor-server/issues/112
Issue: #29
Issue: #27